requirements.txt
rdflib==6.1.1
requests==2.24.0
python-dotenv==0.20.0
schedule==1.1.0
PyGithub==1.55.0